Give physically weak PCs free daily gym sessions
PCs with a physical attribute sum (STR + CON + DEX + AGI) of less than 50 are
given one free training session each day that they visit the gym.
This is based on the assumption (as stated in the help files) that an average
person's stats would be in the 9-12 range. Once the PC is noticeably more
physically fit than average, they are no longer given special treatment.

Also tweaked logic that decides which message to display when leaving the gym.