Free daily training for players with <13 power
glob.h -
Keep track of last gifted day and available session

guild2.cpp -
Logic for gifting sessions and gaining power

omega.cpp -
Set initial values for new globals

save.cpp -
Add new globals to savefile

Bugfix -
It was possible to save your game, restore, and get another 2000 Au
credit for the gym on the same day. We're now saving LastFreeGymDay
global to the savefile so that this cannot be abused.